+++ lib/amazonka-core/src/Amazonka/Data/Path.hs
@@ -10,6 +10,7 @@ module Amazonka.Data.Path
     Path (..),
     RawPath,
     EscapedPath,
+    TwiceEscapedPath,
 
     -- * Constructing Paths
     ToPath (..),
@@ -17,6 +18,7 @@ module Amazonka.Data.Path
 
     -- * Manipulating Paths
     escapePath,
+    escapePathTwice,
     collapsePath,
   )
 where
@@ -59,6 +61,9 @@ type RawPath = Path 'NoEncoding
 
 type EscapedPath = Path 'Percent
 
+newtype TwiceEscapedPath = TwiceEscapedPath (Path 'Percent)
+  deriving newtype (Eq, Show, ToByteString)
+
 instance Semigroup RawPath where
   Raw xs <> Raw ys = Raw (xs ++ ys)
 
@@ -73,6 +78,13 @@ instance ToByteString EscapedPath where
 escapePath :: Path a -> EscapedPath
 escapePath (Raw xs) = Encoded (map (URI.urlEncode True) xs)
 escapePath (Encoded xs) = Encoded xs
+
+-- | Escape a path twice for rare occasions like computing the sigv4 canonical path.
+escapePathTwice :: Path a -> TwiceEscapedPath
+escapePathTwice p = TwiceEscapedPath $
+  Encoded $ case p of
+    Raw xs -> map (URI.urlEncode True . URI.urlEncode True) xs
+    Encoded xs -> map (URI.urlEncode True) xs
 
 collapsePath :: Path a -> Path a
 collapsePath = \case