+++ lib/amazonka-core/src/Amazonka/Sign/V4/Base.hs
@@ -30,6 +30,7 @@ data V4 = V4
     metaPath :: Path,
     metaEndpoint :: Endpoint,
     metaCredential :: Credential,
+    metaCanonicalPath :: CanonicalPath,
     metaCanonicalQuery :: CanonicalQuery,
     metaCanonicalRequest :: CanonicalRequest,
     metaCanonicalHeaders :: CanonicalHeaders,
@@ -125,6 +126,8 @@ type NormalisedHeaders = Tag "normalised-headers" [(ByteString, ByteString)]
 
 type Method = Tag "method" ByteString
 
+type CanonicalPath = Tag "canonical-path" ByteString
+
 type Path = Tag "path" ByteString
 
 type Signature = Tag "signature" ByteString
@@ -181,6 +184,7 @@ signMetadata a r ts presign digest rq =
       metaPath = path,
       metaEndpoint = end,
       metaCredential = cred,
+      metaCanonicalPath = cpath,
       metaCanonicalQuery = query,
       metaCanonicalRequest = crq,
       metaCanonicalHeaders = chs,
@@ -196,7 +200,7 @@ signMetadata a r ts presign digest rq =
     sts = stringToSign ts scope crq
     cred = credential (_authAccessKeyId a) scope
     scope = credentialScope svc end ts
-    crq = canonicalRequest method path digest query chs shs
+    crq = canonicalRequest method cpath digest query chs shs
 
     chs = canonicalHeaders headers
     shs = signedHeaders headers
@@ -204,7 +208,7 @@ signMetadata a r ts presign digest rq =
 
     end = _serviceEndpoint svc r
     method = Tag . toBS $ _requestMethod rq
-    path = escapedPath rq
+    (path, cpath) = escapedPaths rq
 
     svc = _requestService rq
 
@@ -243,7 +247,7 @@ credentialScope s e t =
 
 canonicalRequest ::
   Method ->
-  Path ->
+  CanonicalPath ->
   Hash ->
   CanonicalQuery ->
   CanonicalHeaders ->
@@ -261,11 +265,15 @@ canonicalRequest meth path digest query chs shs =
         toBS digest
       ]
 
-escapedPath :: Request a -> Path
-escapedPath r = Tag . toBS . escapePath $
-  case _serviceAbbrev (_requestService r) of
-    "S3" -> _requestPath r
-    _ -> collapsePath (_requestPath r)
+escapedPaths :: Request a -> (Path, CanonicalPath)
+escapedPaths r = case _serviceAbbrev (_requestService r) of
+  "S3" -> (Tag escapedPath, Tag escapedPath)
+  _ -> (Tag (toBS escapedCollapsed), Tag (toBS canonicalPath))
+  where
+    path = _requestPath r
+    escapedPath = toBS (escapePath path)
+    escapedCollapsed = escapePath (collapsePath path)
+    canonicalPath = escapePathTwice escapedCollapsed
 
 canonicalQuery :: QueryString -> CanonicalQuery
 canonicalQuery = Tag . toBS